# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ability-advocacy  API Documentation",
    "description": "API documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": "http",
  "paths": {
...
    "/kid/add-kid": {
      "post": {
        "tags": [
          "Kids"
        ],
        "summary": "add-kid",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "date_of_birth",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "diagnosis",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "attachments",
            "in": "formData",
            "type": "array",
            "items": {
              "type": "file"
            }
          },
          {
            "name": "profile_image",
            "in": "formData",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/{id}": {
      "get": {
        "tags": [
          "Kids"
        ],
        "summary": "get-profile",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/change-status": {
      "patch": {
        "tags": [
          "Kids"
        ],
        "summary": "status-change",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/statusSwaggerSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/remove-kid/{id}": {
      "put": {
        "tags": [
          "Kids"
        ],
        "summary": "remove-kid",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/update-kid": {
      "put": {
        "tags": [
          "Kids"
        ],
        "summary": "update-Kid",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "kidId",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "date_of_birth",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "diagnosis",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "attachments",
            "in": "formData",
            "type": "array",
            "items": {
              "type": "file"
            }
          },
          {
            "name": "removeAttachments",
            "in": "formData",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "profile_image",
            "in": "formData",
            "type": "file"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "removeAttachments": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/milestones/submit-answer": {
      "post": {
        "tags": [
          "Kids"
        ],
        "summary": "submit-milestone-answer",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/submitMilestoneSwaggerSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/milestones/pending-questions/{kidId}": {
      "get": {
        "tags": [
          "Kids"
        ],
        "summary": "pending-questions-list",
        "description": "",
        "parameters": [
          {
            "name": "kidId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kid/milestones/activity/{kidId}": {
      "get": {
        "tags": [
          "Kids"
        ],
        "summary": "milestone-activity-list",
        "description": "",
        "parameters": [
          {
            "name": "kidId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/chat/update-socket-id": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "update-socket-id",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/check-group": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "check-group",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/create-private-group": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "create-private-group",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/create-team-group": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "create-team-group",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "groupName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "members",
            "in": "formData",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "group_icon",
            "in": "formData",
            "type": "file"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "members": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/reset-unread-count": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "reset-unread-count",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/fetch-all-group": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "fetch-all-group",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/get-group-members": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "get-group-members",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/frequently-contacted-list": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "frequently-contacted-list",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/chat/single-group-details/{groupId}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "single-group-details",
        "description": "",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "statusSwaggerSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended"
            ]
          }
        },
        "required": [
          "id",
          "status"
        ],
        "additionalProperties": false
      },
      "submitMilestoneSwaggerSchema": {
        "type": "object",
        "properties": {
          "kidId": {
            "type": "number",
            "format": "float"
          },
          "questionId": {
            "type": "number",
            "format": "float"
          },
          "answer": {
            "type": "boolean"
          }
        },
        "required": [
          "kidId",
          "questionId",
          "answer"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}